refactor(broker): move job stream classes out of JobStreamServiceStep
